Most trivial player test evar. Rewired code so loading JSON isn't implicitly part of `config.py` Merged PyTest files for GameObject Made a mess in main.py and test_player.py in terms of where config loading goes
