+++ furusiya/function_mess.py
@@ -2,7 +2,7 @@ import colors
 import config
 from constants import *
 from main_interface import Game, menu, message, is_blocked
-from model.ai import BasicMonster, ConfusedMonster
+from model.ai import ConfusedMonster
 from model.fighter import Fighter
 from model.game_object import GameObject
 from model.item import Item
