--- tests/furusiya/model/gameobject/conftest.py
@@ -1,27 +1,0 @@
-import pytest
-
-from model.game_object import GameObject
-
-
-@pytest.fixture
-def obj():
-    yield GameObject(1, 1, 'j', 'test', (234, 12, 213))
-
-
-class TestComp:
-    def __init__(self, value=0):
-        self.value = value
-
-
-class TestSubComp(TestComp):
-    pass
-
-
-@pytest.fixture
-def comp():
-    yield TestComp
-
-
-@pytest.fixture
-def subcomp():
-    yield TestSubComp
